Day_9/smart-tax-compliance-tracker/requirements.txt
streamlit
langchain
langgraph
faiss-cpu
chromadb
sentence-transformers
PyPDF2
python-docx
openpyxl
icalendar
python-dotenv
google-generativeai
